move cluster sql tests
