--- packages/effect/test/cluster/fixtures/mysql2-utils.ts
@@ -1,1 +1,0 @@
-../../../../sql/mysql2/test/utils.ts
