--- packages/effect/test/cluster/fixtures/pg-utils.ts
@@ -1,1 +1,0 @@
-../../../../sql/pg/test/utils.ts
