+++ packages/effect/tsconfig.test.json
@@ -2,13 +2,7 @@
   "$schema": "http://json.schemastore.org/tsconfig",
   "extends": "../../tsconfig.base.jsonc",
   "include": ["test", "dtslint", "benchmark"],
-  "references": [
-    { "path": "tsconfig.src.json" },
-    { "path": "../platform-node/tsconfig.src.json" },
-    { "path": "../sql/pg/tsconfig.src.json" },
-    { "path": "../sql/mysql2/tsconfig.src.json" },
-    { "path": "../sql/sqlite-node/tsconfig.src.json" }
-  ],
+  "references": [{ "path": "tsconfig.src.json" }],
   "compilerOptions": {
     "tsBuildInfoFile": ".tsbuildinfo/test.tsbuildinfo",
     "outDir": ".tsbuildinfo/test",
@@ -17,15 +11,7 @@
     "baseUrl": ".",
     "paths": {
       "effect": ["src/index.ts"],
-      "effect/*": ["src/*/index.ts", "src/*.ts"],
-      "@effect/platform-node": ["../platform-node/src/index.ts"],
-      "@effect/platform-node/*": ["../platform-node/src/*/index.ts"],
-      "@effect/sql-mysql2": ["../sql/mysql2/src/index.ts"],
-      "@effect/sql-mysql2/*": ["../sql/mysql2/src/*/index.ts"],
-      "@effect/sql-pg": ["../sql/pg/src/index.ts"],
-      "@effect/sql-pg/*": ["../sql/pg/src/*/index.ts"],
-      "@effect/sql-sqlite-node": ["../sql/sqlite-node/src/index.ts"],
-      "@effect/sql-sqlite-node/*": ["../sql/sqlite-node/src/*/index.ts"]
+      "effect/*": ["src/*/index.ts", "src/*.ts"]
     },
     "plugins": [
       {
