+++ packages/platform-node/test/cluster/SqlMessageStorage.test.ts
@@ -3,11 +3,6 @@ import { SqliteClient } from "@effect/sql-sqlite-node"
 import { describe, expect, it } from "@effect/vitest"
 import { Effect, Fiber, Layer } from "effect"
 import { FileSystem } from "effect/platform"
-import { TestClock } from "effect/testing"
-import { Message, MessageStorage, ShardingConfig, Snowflake, SqlMessageStorage } from "effect/unstable/cluster"
-import { SqlClient } from "effect/unstable/sql"
-import { MysqlContainer } from "./fixtures/mysql2-utils.ts"
-import { PgContainer } from "./fixtures/pg-utils.ts"
 import {
   makeAckChunk,
   makeChunkReply,
@@ -15,7 +10,12 @@ import {
   makeRequest,
   PrimaryKeyTest,
   StreamRpc
-} from "./MessageStorage.test.js"
+} from "effect/test/cluster/MessageStorage.test"
+import { TestClock } from "effect/testing"
+import { Message, MessageStorage, ShardingConfig, Snowflake, SqlMessageStorage } from "effect/unstable/cluster"
+import { SqlClient } from "effect/unstable/sql"
+import { MysqlContainer } from "../fixtures/mysql2-utils.ts"
+import { PgContainer } from "../fixtures/pg-utils.ts"
 
 const StorageLive = SqlMessageStorage.layer.pipe(
   Layer.provideMerge(Snowflake.layerGenerator),
