+++ packages/platform-node/test/fixtures/mysql2-utils.ts
@@ -1,0 +1,1 @@
+../../../sql/mysql2/test/utils.ts
