+++ packages/platform-node/test/fixtures/pg-utils.ts
@@ -1,0 +1,1 @@
+../../../sql/pg/test/utils.ts
