+++ packages/platform-node/tsconfig.test.json
@@ -2,7 +2,13 @@
   "$schema": "http://json.schemastore.org/tsconfig",
   "extends": "../../tsconfig.base.jsonc",
   "include": ["test/**/*.ts", "test/**/*.json"],
-  "references": [{ "path": "tsconfig.src.json" }],
+  "references": [
+    { "path": "tsconfig.src.json" },
+    { "path": "../effect/tsconfig.test.json" },
+    { "path": "../sql/pg/tsconfig.src.json" },
+    { "path": "../sql/mysql2/tsconfig.src.json" },
+    { "path": "../sql/sqlite-node/tsconfig.src.json" }
+  ],
   "compilerOptions": {
     "tsBuildInfoFile": ".tsbuildinfo/test.tsbuildinfo",
     "outDir": ".tsbuildinfo/test",
@@ -13,7 +19,14 @@
     "baseUrl": ".",
     "paths": {
       "@effect/platform-node": ["src/index.ts"],
-      "@effect/platform-node/*": ["src/*/index.ts", "src/*.ts"]
+      "@effect/platform-node/*": ["src/*/index.ts", "src/*.ts"],
+      "effect/test/*": ["../effect/test/*.ts"],
+      "@effect/sql-mysql2": ["../sql/mysql2/src/index.ts"],
+      "@effect/sql-mysql2/*": ["../sql/mysql2/src/*.ts"],
+      "@effect/sql-pg": ["../sql/pg/src/index.ts"],
+      "@effect/sql-pg/*": ["../sql/pg/src/*.ts"],
+      "@effect/sql-sqlite-node": ["../sql/sqlite-node/src/index.ts"],
+      "@effect/sql-sqlite-node/*": ["../sql/sqlite-node/src/*.ts"]
     },
     "plugins": [
       {
